requirements.txt
# packaging
pip>=9
setuptools>=26
twine
wheel>=0.29

# data science
numpy
pandas
scikit-learn >=1.2.2
seaborn
plotly
pycountry

# nlp
#gensim
#spacy

# UI
#dash
#dash_bootstrap_components
#streamlir

# tests/linter
black

# coverage
# flake8
# pytest
# yapf

# API
# gcsfs
# google-cloud-storage
# mlflow
# s3fs

# utilities
# six>=1.14
joblib
pyyaml>=6.0
# memoized-property
# termcolor
